requirements.txt
networkx==2.4
numpy==1.21.3
sympy==1.10.1
openai==0.27.8
spacy==3.0.5
nltk==3.4.5
tqdm==4.42.1
transformers==3.5.1
graphviz==0.20.1
pandas==1.3.4
tenacity==8.2.2
benepar==0.2.0
jsonlines==3.1.0
python-dotenv==1.0.0
scikit_learn==1.2.2
